backend/requirements.txt
Flask==2.3.2
opencv-python==4.8.0.76
numpy==1.25.2
scikit-learn==1.3.0
Pillow==10.1.0
